+++ numba_xnd/__init__.py
@@ -1,0 +1,3 @@
+# register libraries
+from . import libndtypes as _  # NOQA
+from . import ndtypes as _  # NOQA
