+++ numba_xnd/libndtypes/api.py
@@ -1,0 +1,10 @@
+def ndt_as_ndarray(a, n, ctx):
+    raise NotImplementedError()
+
+
+def create_ndt_ndarray():
+    raise NotImplementedError()
+
+
+def create_ndt_context():
+    raise NotImplementedError()
