+++ numba_xnd/ndtypes/__init__.py
@@ -1,0 +1,10 @@
+# register types
+from . import typing as _  # NOQA
+
+# register lowering
+from . import lowering as _  # NOQA
+
+# expose
+from .api import *  # NOQA
+from .models import *  # NOQA
+from .types import *  # NOQA
