+++ numba_xnd/ndtypes/api.py
@@ -1,0 +1,6 @@
+def py_ndt_to_ndt(x):
+    raise NotImplementedError()
+
+
+def ndt_to_py_ndt(x):
+    raise NotImplementedError()
