+++ numba_xnd/ndtypes/models.py
@@ -1,0 +1,6 @@
+from numba.extending import register_model
+from ..libndtypes import NdtModel
+
+from .types import PyNdtType
+
+register_model(PyNdtType)(NdtModel)
