+++ numba_xnd/shared/__init__.py
@@ -1,0 +1,3 @@
+# expose
+from .llvm import *  # NOQA
+from .sizes import *  # NOQA
