+++ numba_xnd/shared/sizes.py
@@ -2,5 +2,6 @@ CONST_NDT_MAX_DIM = 128
 SIZEOF_NDT_T = 112
 SIZEOF_NDT_NDARRAY_T = 3088
 SIZEOF_NDT_CONTEXT_T = 24
+OFFSETOF_NDT_T_NDIM = 12
 OFFSETOF_NDT_ARRAY_T_NDIM = 0
 OFFSETOF_NDT_ARRAY_T_SHAPE = 16
