+++ tests/test_ndtypes.py
@@ -3,8 +3,7 @@ import unittest
 from numba import njit
 from ndtypes import ndt
 
-# register numba things
-import numba_xnd.ndtypes.python
+import numba_xnd
 
 
 n = ndt("10 * 4 * 4 * int64")
@@ -16,7 +15,6 @@ class TestPyNdt(unittest.TestCase):
 
     def test_ndim(self):
         self.assertEqual(njit(lambda x: x.ndim)(n), 3)
-
 
     def test_shape(self):
         self.assertSequenceEqual(njit(lambda x: x.shape)(n), [10, 4, 4])
